fix random player size, add OnPlayerReset

+++ desktop_version/src/Graphics.h
@@ -150,15 +150,15 @@ public:
 
     void draw_texture(SDL_Texture* image, int x, int y);
 
-    void draw_texture_part(SDL_Texture* image, int x, int y, int x2, int y2, int w, int h, int scalex, int scaley);
+    void draw_texture_part(SDL_Texture* image, int x, int y, int x2, int y2, int w, int h, float scalex, float scaley);
 
-    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int scalex, int scaley);
+    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, float scalex, float scaley);
     void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height);
-    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b, int a, int scalex, int scaley);
+    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b, int a, float scalex, float scaley);
     void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b, int a);
-    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b, int scalex, int scaley);
+    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b, float scalex, float scaley);
     void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, int r, int g, int b);
-    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, SDL_Color color, int scalex, int scaley);
+    void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, SDL_Color color, float scalex, float scaley);
     void draw_grid_tile(SDL_Texture* texture, int t, int x, int y, int width, int height, SDL_Color color);
 
     void updatetextboxes(void);
